Handled some overflow in Java

+++ java/generic/IPrimitiveRoots.java
@@ -1,7 +1,7 @@
 package generic;
 
 interface IPrimitiveRoots<T extends IField<T>> {
-    T primitiveRoot(int n);
-    T pow(int exp);
+    T primitiveRoot(long n);
+    T pow(long exp);
     T[] precomputeRootsOfUnity(int n, int direction);
 }
